replace APIController with Api model
