+++ src/contexts/Api/index.tsx
@@ -29,12 +29,12 @@ import {
   defaultNetworkMetrics,
   defaultStakingMetrics,
 } from './defaults';
-import { APIController } from 'static/APIController';
 import { isCustomEvent } from 'static/utils';
-import type { ApiStatus } from 'static/APIController/types';
 import { useEventListener } from 'usehooks-ts';
 import BigNumber from 'bignumber.js';
 import { SyncController } from 'static/SyncController';
+import { ApiController } from 'controllers/ApiController';
+import type { ApiStatus, ConnectionType } from 'model/Api/types';
 
 export const APIContext = createContext<APIContextInterface>(defaultApiContext);
 
@@ -48,6 +48,9 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
   const [isLightClient, setIsLightClientState] = useState<boolean>(
     !!localStorage.getItem('light_client')
   );
+
+  // Whether this context has initialised.
+  const initialisedRef = useRef<boolean>(false);
 
   // Setter for whether light client is active. Updates state and local storage.
   const setIsLightClient = (value: boolean) => {
@@ -115,8 +118,7 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
 
   // Fetch chain state. Called once `provider` has been initialised.
   const onApiReady = async () => {
-    // TODO: get the correct `api` instance from `ApiController`.
-    const { api } = APIController;
+    const { api } = ApiController.get(network);
 
     const newChainState = await Promise.all([
       api.rpc.system.chain(),
@@ -130,7 +132,6 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
       const chain = newChainState[0]?.toString();
       const version = newChainState[1]?.toJSON();
       const ss58Prefix = Number(newChainState[2]?.toString());
-
       setChainState({ chain, version, ss58Prefix });
     }
 
@@ -140,14 +141,15 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
 
   // Bootstrap app-wide chain state.
   const bootstrapNetworkConfig = async () => {
-    // TODO: Bootstrap Api instance from ApiController.
+    const apiInstance = ApiController.get(network);
+
     const {
       consts: newConsts,
       networkMetrics: newNetworkMetrics,
       activeEra: newActiveEra,
       poolsConfig: newPoolsConfig,
       stakingMetrics: newStakingMetrics,
-    } = await APIController.bootstrapNetworkConfig();
+    } = await apiInstance.bootstrapNetworkConfig();
 
     // Populate all config state.
     setConsts(newConsts);
@@ -168,10 +170,9 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
     SyncController.dispatch('initialization', 'complete');
 
     // Initialise subscriptions.
-    // TODO: initialise all this from ApiController.
-    APIController.subscribeNetworkMetrics();
-    APIController.subscribePoolsConfig();
-    APIController.subscribeActiveEra();
+    apiInstance.subscribeNetworkMetrics();
+    apiInstance.subscribePoolsConfig();
+    apiInstance.subscribeActiveEra();
   };
 
   // Handle Api disconnection.
@@ -311,96 +312,78 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
     );
   };
 
-  // Handle an initial api connection.
-  useEffect(() => {
-    // NOTE: `provider` should always be present after the first call to initialize. We therefore
-    // can check whether on initial api connection based  `provider`.
+  const reInitialiseApi = async (type: ConnectionType) => {
+    await ApiController.instantiate(network, type, rpcEndpoint);
+  };
 
-    // TODO: Use an initialisation ref to check whether this is the first call to initialize, and initialize an Api instance for the network.
-    if (!APIController.provider) {
-      APIController.initialize(
-        network,
-        isLightClient ? 'sc' : 'ws',
-        rpcEndpoint,
-        {
-          clearState: true,
-          initial: true,
-        }
-      );
+  // Handle initial api connection.
+  useEffect(() => {
+    // Uses initialisation ref to check whether this is the first context render, and initializes an Api instance for the current network if that is the case.
+    if (!initialisedRef.current) {
+      initialisedRef.current = true;
+      reInitialiseApi(isLightClient ? 'sc' : 'ws');
     }
   });
 
-  // If RPC endpoint changes, and not on light client, re-connect.
-  useEffectIgnoreInitial(() => {
+  // If RPC endpoint changes, and not on light client, re-initialise API.
+  useEffectIgnoreInitial(async () => {
     if (!isLightClient) {
-      // TODO: use instance via ApiController to destroy the current instance and create a new one.
-      APIController.initialize(network, 'ws', rpcEndpoint, {
-        clearState: false,
-      });
+      reInitialiseApi('ws');
     }
   }, [rpcEndpoint]);
 
-  // Trigger API reconnect on network or light client change.
+  // If connection type changes, re-initialise API.
+  useEffectIgnoreInitial(async () => {
+    const apiInstance = ApiController.get(network);
+    if (apiInstance.connectionType === 'sc') {
+      reInitialiseApi(isLightClient ? 'sc' : 'ws');
+    }
+  }, [isLightClient]);
+
+  // Re-initialise API and set defaults on network change.
   useEffectIgnoreInitial(() => {
     setRpcEndpoint(initialRpcEndpoint());
 
-    // TODO: instead of checking network change, check if the current Api instance is of the `sc`
-    // type. If so, do not clear state.
-    const networkSwitch = network !== APIController.network;
-
-    // If network changes, reset consts and chain state.
-    if (networkSwitch) {
-      setConsts(defaultConsts);
-      setChainState(defaultChainState);
-      setStateWithRef(
-        defaultNetworkMetrics,
-        setNetworkMetrics,
-        networkMetricsRef
-      );
-      setStateWithRef(defaultActiveEra, setActiveEra, activeEraRef);
-      setStateWithRef(defaultPoolsConfig, setPoolsConfig, poolsConfigRef);
-      setStateWithRef(
-        defaultStakingMetrics,
-        setStakingMetrics,
-        stakingMetricsRef
-      );
-    }
-
-    // Reconnect API instance, clearing state only if network has changed.
-    // TODO: use instance via ApiController to destroy the current instance and create a new one.
-    APIController.initialize(
-      network,
-      isLightClient ? 'sc' : 'ws',
-      rpcEndpoint,
-      { clearState: !networkSwitch }
+    // Reset consts and chain state.
+    setConsts(defaultConsts);
+    setChainState(defaultChainState);
+    setStateWithRef(
+      defaultNetworkMetrics,
+      setNetworkMetrics,
+      networkMetricsRef
     );
-  }, [isLightClient, network]);
+    setStateWithRef(defaultActiveEra, setActiveEra, activeEraRef);
+    setStateWithRef(defaultPoolsConfig, setPoolsConfig, poolsConfigRef);
+    setStateWithRef(
+      defaultStakingMetrics,
+      setStakingMetrics,
+      stakingMetricsRef
+    );
+    setApiStatus('disconnected');
 
-  // Add event listener for `polkadot-api` notifications. Also handles unmounting logic. TODO: Move
-  // this to `useEventListener` hook, and use a separate `useEffect` to call `cancelFn` and
-  // `unsubscribe` on all instnaces.
-  useEffect(() => {
-    document.addEventListener('polkadot-api', handleNewApiStatus);
-    return () => {
-      document.removeEventListener('polkadot-api', handleNewApiStatus);
-      APIController.cancelFn?.();
-      APIController.unsubscribe();
-    };
-  }, []);
+    reInitialiseApi(isLightClient ? 'sc' : 'ws');
+  }, [network]);
 
-  // Add event listener for subscription updates.
+  // Call `cancelFn` and `unsubscribe` on active instnace on unmount.
+  useEffect(
+    () => () => {
+      const instance = ApiController.get(network);
+      instance?.cancelFn?.();
+      instance?.unsubscribe();
+    },
+    []
+  );
+
+  // Add event listener for api events and subscription updates.
   const documentRef = useRef<Document>(document);
-
+  useEventListener('polkadot-api', handleNewApiStatus, documentRef);
   useEventListener(
     'new-network-metrics',
     handleNetworkMetricsUpdate,
     documentRef
   );
-
   useEventListener('new-active-era', handleActiveEraUpdate, documentRef);
-
   useEventListener('new-pools-config', handlePoolsConfigUpdate, documentRef);
-
   useEventListener(
     'new-staking-metrics',
     handleStakingMetricsUpdate,
@@ -410,8 +393,7 @@ export const APIProvider = ({ children, network }: APIProviderProps) => {
   return (
     <APIContext.Provider
       value={{
-        // TODO: use correct instance from ApiController.
-        api: APIController.api,
+        api: ApiController.get(network)?.api || null,
         chainState,
         apiStatus,
         isLightClient,
