+++ src/contexts/Api/types.ts
@@ -6,7 +6,7 @@ import type { U8aLike } from '@polkadot/util/types';
 import type BigNumber from 'bignumber.js';
 import type { ReactNode } from 'react';
 import type { AnyJson, NetworkName } from '../../types';
-import type { ApiStatus } from 'static/APIController/types';
+import type { ApiStatus } from 'model/Api/types';
 
 export interface APIProviderProps {
   children: ReactNode;
