+++ src/contexts/Connect/ImportedAccounts/index.tsx
@@ -29,7 +29,7 @@ export const ImportedAccountsProvider = ({
 }: {
   children: ReactNode;
 }) => {
-  const { isReady } = useApi();
+  const { isReady, api } = useApi();
   const { otherAccounts } = useOtherAccounts();
   const { extensionAccounts } = useExtensionAccounts();
   const allAccounts = extensionAccounts.concat(otherAccounts);
@@ -104,8 +104,11 @@ export const ImportedAccountsProvider = ({
 
   // Keep accounts in sync with `BalancesController`.
   useEffectIgnoreInitial(() => {
-    if (isReady) {
-      BalancesController.syncAccounts(allAccounts.map((a) => a.address));
+    if (api && isReady) {
+      BalancesController.syncAccounts(
+        api,
+        allAccounts.map((a) => a.address)
+      );
     }
   }, [isReady, allAccountsStringified]);
