+++ src/model/Api/types.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { ConnectionType, EventStatus } from 'static/APIController/types';
+import type * as ScType from '@substrate/connect';
 import type { NetworkName } from 'types';
 
 export interface APIConfig {
@@ -15,3 +15,14 @@ export interface APIEventDetail {
   network: NetworkName;
   err?: string;
 }
+
+export interface SubstrateConnect {
+  WellKnownChain: (typeof ScType)['WellKnownChain'];
+  createScClient: (typeof ScType)['createScClient'];
+}
+
+export type ConnectionType = 'ws' | 'sc';
+
+export type ApiStatus = 'connecting' | 'connected' | 'disconnected' | 'ready';
+
+export type EventStatus = ApiStatus | 'error';
