+++ src/static/ActivePoolsController/index.ts
@@ -4,12 +4,12 @@
 import type { VoidFn } from '@polkadot/api/types';
 import { defaultPoolNominations } from 'contexts/Pools/ActivePool/defaults';
 import type { ActivePool, PoolRoles } from 'contexts/Pools/ActivePool/types';
-import { APIController } from 'static/APIController';
 import { IdentitiesController } from 'static/IdentitiesController';
 import type { AnyApi } from 'types';
 import type { ActivePoolItem, DetailActivePool } from './types';
 import { SyncController } from 'static/SyncController';
 import type { Nominations } from 'contexts/Balances/types';
+import type { ApiPromise } from '@polkadot/api';
 
 export class ActivePoolsController {
   // ------------------------------------------------------
@@ -33,9 +33,10 @@ export class ActivePoolsController {
   // ------------------------------------------------------
 
   // Subscribes to pools and unsubscribes from removed pools.
-  static syncPools = async (newPools: ActivePoolItem[]): Promise<void> => {
-    const { api } = APIController;
-
+  static syncPools = async (
+    api: ApiPromise,
+    newPools: ActivePoolItem[]
+  ): Promise<void> => {
     // Sync: Checking active pools.
     SyncController.dispatch('active-pools', 'syncing');
 
@@ -67,6 +68,7 @@ export class ActivePoolsController {
           ]): Promise<void> => {
             // NOTE: async: fetches identity data for roles.
             await this.handleActivePoolCallback(
+              api,
               pool,
               bondedPool,
               rewardPool,
@@ -96,6 +98,7 @@ export class ActivePoolsController {
 
   // Handle active pool callback.
   static handleActivePoolCallback = async (
+    api: ApiPromise,
     pool: ActivePoolItem,
     bondedPoolResult: AnyApi,
     rewardPoolResult: AnyApi,
@@ -108,6 +111,7 @@ export class ActivePoolsController {
 
     // Fetch identities for roles and expand `bondedPool` state to store them.
     bondedPool.roleIdentities = await IdentitiesController.fetch(
+      api,
       this.getUniqueRoleAddresses(bondedPool.roles)
     );
